build: Add dependency on GLEW to perform OpenGL loading
TR1X previously contained static code generated by a GLAD-style tool
to perform OpenGL loading, which has the downside that it needs to be
manually maintained. This had not been done, and so the existing GL
loading code e.g. only supported GLX, but not the modern EGL interface.

This change outsources GL loading to the very commonly used GLEW
library instead.